i18n
Implemented i18n, only english so far. Other languages can be implemented in the future.

Still requires testing with another language to make sure all texts are covered.

Next commit will be testing with Portuguese language.

It's also necessary to add a selector in the settings screen to allow changing between languages and potentially an automatic detector at launch to startup with the device's language if possible.